backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.30
psycopg[binary]==3.1.19
alembic==1.13.2
python-jose==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
pydantic-settings==2.3.4
httpx==0.27.0
qdrant-client==1.9.1
boto3==1.34.162
openai==1.35.3
tenacity==8.3.0
